only snapshot the last 30000 blocks
